requirements.txt
fastapi[all]
sqlalchemy
pyodbc
python-dotenv
python-jose[cryptography]
passlib[bcrypt]